Refactor: change naming without fist letter to change the lowerCamelCase to UpperCamelCase
