+++ src/routes/outes.tsx
@@ -1,5 +1,5 @@
 import { createBrowserRouter } from 'react-router-dom';
-import { AuthRoutes } from './AuthRoutes';
-import { RootRoutes } from './RootRoutes';
+import { AuthRoutes } from './uthRoutes';
+import { RootRoutes } from './ootRoutes';
 
 export const Routes = createBrowserRouter([AuthRoutes, RootRoutes]);
